+++ src/main/java/com/maan/eway/bean/SectionCoverMaster.java
@@ -19,6 +19,7 @@ import org.hibernate.annotations.DynamicUpdate;
 import lombok.*;
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 
 import java.util.Date;
@@ -125,7 +126,7 @@ private static final long serialVersionUID = 1L;
     private String     toolTip ;
 
     @Column(name="COVERAGE_LIMIT")
-    private Double     coverageLimit ;
+    private BigDecimal     coverageLimit ;
 
     @Column(name="CALC_TYPE", length=1)
     private String     calcType ;
@@ -134,22 +135,22 @@ private static final long serialVersionUID = 1L;
     private String     calcTypeDesc ;
 
     @Column(name="BASE_RATE")
-    private Double     baseRate ;
+    private BigDecimal     baseRate ;
 
     @Column(name="EXCESS")
-    private Double     excess ;
+    private BigDecimal     excess ;
 
     @Column(name="MIN_PREMIUM")
-    private Double     minPremium ;
+    private BigDecimal     minPremium ;
 
     @Column(name="MIN_SUMINSURED")
-    private Double     minSuminsured ;
+    private BigDecimal     minSuminsured ;
 
     @Column(name="CREATED_BY", length=100)
     private String     createdBy ;
 
     @Column(name="MAX_SUMINSURED")
-    private Double     maxSuminsured ;
+    private BigDecimal     maxSuminsured ;
 
     @Column(name="FACTOR_TYPE_ID")
     private Integer    factorTypeId ;
@@ -161,7 +162,7 @@ private static final long serialVersionUID = 1L;
     private String     regulatoryCode ;
 
     @Column(name="TAX_AMOUNT")
-    private Double     taxAmount ;
+    private BigDecimal     taxAmount ;
 
     @Column(name="TAX_CODE", length=100)
     private String     taxCode ;
